+++ cpp11armadillotest/src/07_typedefs.cpp
@@ -26,3 +26,13 @@
   Mat<int> y = as_Mat(x);
   return as_integers_matrix(y);
 }
+
+[[cpp11::register]] doubles_matrix<> typedef_SpMat_double(const doubles_matrix<>& x) {
+  SpMat<double> y = as_SpMat(x);
+  return as_doubles_matrix(y);
+}
+
+[[cpp11::register]] integers_matrix<> typedef_SpMat_int(const integers_matrix<>& x) {
+  SpMat<int> y = as_SpMat(x);
+  return as_integers_matrix(y);
+}